# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/order/addcoupon": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.Coupon"
                ],
                "summary": "api for admin to add coupon",
                "operationId": "AddCoupon",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Coupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "coupon added successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/order/editcoupon": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "edit coupon",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin.Coupon"
                ],
                "summary": "API for edit coupon by admin",
                "operationId": "EditCoupon",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EditCoupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully address added"
                    },
                    "400": {
                        "description": "string \"Invalid input"
                    }
                }
            }
        },
        "/admin/order/updaterderstatus": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.OrderDash"
                ],
                "summary": "api for update order status of a user/order by using id",
                "operationId": "UpdateOrderStatus",
                "parameters": [
                    {
                        "description": "Input Fields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpDateOrderStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated order status"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
...
        "/user/cart/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get cart list pf autherised user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User.Cart"
                ],
                "summary": "API for get cart list",
                "operationId": "ListCart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ListCartResponse"
                        }
                    },
                    "400": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/user/coupon": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.Coupon",
                    "User"
                ],
                "summary": "api for get coupons for admin and user",
                "operationId": "GetCoupon",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Coupon"
                        }
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
...
        "/user/order/checkout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "place order",
                "tags": [
                    "User.Order"
                ],
                "summary": "API for order cart products",
                "operationId": "OrderCartProducts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "coupon",
                        "name": "coupon",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pay method",
                        "name": "payMethod",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "html added",
                        "schema": {
                            "$ref": "#/definitions/response.Order"
                        }
                    },
                    "400": {
                        "description": "somthing wrong!!"
                    },
                    "406": {
                        "description": "string \"Invalid input"
                    }
                }
            }
        },
        "/user/order/confirm-details": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User.Order"
                ],
                "summary": "API for checking all details for buying cart products",
                "operationId": "BuyProduct",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CheckoutResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    }
                }
            }
        },
        "/user/order/invoice/:orderID": {
            "get": {
                "description": "Download the invoice as a PDF file.",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "User.Order"
                ],
                "summary": "Download invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "faild"
                    },
                    "500": {
                        "description": "faild"
                    }
                }
            }
        },
        "/user/order/orderdetails": {
            "get": {
                "description": "order details get by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User.Order"
                ],
                "summary": "order details get by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Orders1"
                        }
                    },
                    "400": {
                        "description": "faild"
                    },
                    "500": {
                        "description": "faild"
                    }
                }
            }
        },
...
        "/user/order/returnorder": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User.Order"
                ],
                "summary": "api for return order",
                "operationId": "ReturnOrder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "orderID",
                        "name": "ID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "requst to return successfully"
                    },
                    "400": {
                        "description": "Missing or invalid entry/error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Address": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "land_mark": {
                    "type": "string"
                },
                "muncipality": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "pin_code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
        "domain.Cart": {
            "type": "object",
            "properties": {
                "Quantity": {
                    "type": "integer"
                },
                "Variation_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Coupon": {
            "type": "object",
            "required": [
                "code",
                "discount_max_amount",
                "discount_percent",
                "min_order_value",
                "valid_days"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 8,
                    "minLength": 4
                },
                "description": {
                    "type": "string"
                },
                "discount_max_amount": {
                    "type": "number"
                },
                "discount_percent": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "min_order_value": {
                    "type": "number"
                },
                "valid_days": {
                    "type": "integer"
                }
            }
        },
...
        "handler.CheckoutResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/domain.Address"
                },
                "cart": {
                    "$ref": "#/definitions/domain.Cart"
                }
            }
        },
        "handler.ListCartResponse": {
            "type": "object",
            "properties": {
                "discountPrice": {
                    "type": "number"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Cart"
                    }
                },
                "totalPrice": {
                    "type": "number"
                }
            }
        },
...
        "request.EditCoupon": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discount_max_amount": {
                    "type": "number"
                },
                "discount_percent": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "min_order_value": {
                    "type": "number"
                },
                "valid_days": {
                    "type": "integer"
                }
            }
        },
...
        "request.UpDateOrderStatus": {
            "type": "object",
            "required": [
                "order_id",
                "status",
                "users_id"
            ],
            "properties": {
                "order_id": {
                    "type": "integer"
                },
                "order_status_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.Cart": {
            "type": "object",
            "properties": {
                "Quantity": {
                    "type": "integer"
                },
                "Variation_name": {
                    "type": "string"
                },
                "actual_price": {
                    "type": "number"
                },
                "discount_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "qty_in_stock": {
                    "type": "integer"
                }
            }
        },
        "response.Order": {
            "type": "object",
            "properties": {
                "Razor_pay_order_id": {
                    "type": "string"
                },
                "actual_price": {
                    "type": "number"
                },
                "disc_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "net_amount": {
                    "type": "number"
                },
                "order_date": {
                    "type": "string"
                },
                "order_status_id": {
                    "type": "integer"
                },
                "payment_id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
        "response.Orders1": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number"
                },
                "coupon_code": {
                    "type": "string"
                },
                "disc_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "description": "Use jsonb type for JSON data",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Cart"
                    }
                },
                "net_amount": {
                    "type": "number"
                },
                "order_date": {
                    "type": "string"
                },
                "order_status_name": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                }
            }
        },
...
    }
}